requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
boto==2.42.0
boto3==1.3.0
botocore==1.4.5
dj-database-url==0.4.0
Django==1.10
django-bmemcached==0.2.3
django-cors-headers==1.1.0
django-filter==0.13.0
django-mongodb-engine==0.6.0
django-redis-cache==1.6.5
django-ses==0.7.0
django-storages==1.5.0
djangorestframework==3.3.3
djangotoolbox==1.8.0
docutils==0.12
enum34==1.1.2
funcsigs==0.4
futures==3.0.5
gunicorn==19.4.5
jmespath==0.9.0
kombu==3.0.35
Markdown==2.6.6
mock==2.0.0
mongoengine==0.10.6
pbr==1.10.0
Pillow==3.1.1
psycopg2==2.6.2
pymongo==3.2.2
python-binary-memcached==0.24.6
python-dateutil==2.5.1
python-http-client==1.2.3
python-memcached==1.57
pytz==2016.6.1
selenium==2.53.1
sendgrid==2.2.1
six==1.10.0
smtpapi==0.3.1
virtualenv==14.0.5
whitenoise==2.0.6
redis==2.10.5
celery==3.1.23
django-celery==3.1.17
django-angular==0.8.4
gitignore==0.0.8
